feat(app): full mouse support — drag selection, double/triple click, right-click menu
Editor pane mouse interaction now covers the full GUI-editor
vocabulary, configurable via [editor].mouse = enabled / click-only
/ disabled.

dispatch.rs::handle_mouse:
- Down(Left) inside editor pane positions the cursor (translating
  screen coordinates through the line-number gutter via the new
  narwhal_tui::gutter_width helper) and arms a drag anchor.
- Drag(Left) extends the selection from the anchor to the current
  buffer position.
- Up(Left) closes the drag.
- Down(Left) within 500ms of the previous click at the same cell
  promotes the click count: double = word select (using a new
  word_bounds_at helper), triple+ = full-line select.
- Down(Middle) inside the editor pastes the clipboard at the click
  position (X-style primary-selection paste).
- Down(Right) inside the editor opens the context menu (Cut / Copy
  / Paste / Select All / Run Selection / Find / Toggle Comment).
  Disabled entries are computed live from selection + clipboard
  state.

Mouse plumbing:
- UiState gains LastClick and MouseDragState (Phase 3 added the
  field declarations; this commit fills them).
- editor_dispatch/context_menu.rs handles menu navigation
  (Up/Down/j/k cycle skipping disabled, Enter/Space accepts, Esc
  closes) and dispatches the action.
- narwhal-tui/widgets/context_menu.rs renders the overlay with
  Clear + Block + Paragraph, clamped to screen bounds.
- The vim and modeless handlers both yield to the context menu
  through a shared check at the top of handle_editor_key.

word_bounds_at lives in dispatch.rs with 6 unit tests covering
middle / boundary / whitespace / underscore / overflow cases.

7 integration tests in tests/mouse_editor.rs exercise the full
pipeline end-to-end through TestBackend renders.

+++ crates/narwhal-app/src/core/state/ui.rs
@@ -23,7 +23,7 @@ use super::{ResultState, SidebarItem, StatusBar, Tab};
 /// from there.
 #[derive(Debug, Clone, Copy)]
 pub struct MouseDragState {
-    pub tab_id: usize,
+    pub tab_id: u64,
     pub anchor: (usize, usize),
 }
